refactor: Papago API 코드 분리
