+++ RF/Services/Chat/ChatService.swift
@@ -49,56 +49,6 @@ class ChatService {
         socketClient.sendJSONForDict(dict: object, toDestination: "\(destination)/\(partyId)")
     }
     
-    func translateMessage(source: String, target: String, text: String, completion: @escaping (String)->()) {
-        
-        let url = "\(Domain.naverApi)\(PapagoApiPath.translation)"
-        let headers: HTTPHeaders = [
-            "Content-Type": "application/x-www-form-urlencoded",
-            "X-Naver-Client-Id": NaverApiHeaders.clientId,
-            "X-Naver-Client-Secret": NaverApiHeaders.clientSecret
-        ]
-        let params: [String: Any] = [
-            "source": source,
-            "target": target,
-            "text": text
-        ]
-        
-        AF.request(url, method: .post, parameters: params, encoding: URLEncoding.default, headers: headers)
-            .validate(statusCode: 200..<201)
-            .responseDecodable(of: TranslationData.self) { response in
-                switch response.result{
-                case .success (let data):
-                    completion(data.message.result.translatedText)
-                case .failure (let error):
-                    print("Translate error!\n\(error)")
-                }
-            }
-    }
-    
-    func detectLanguage(_ text: String, completion: @escaping (String)->()) {
-        let url = "\(Domain.naverApi)\(PapagoApiPath.detectLangs)"
-        let headers: HTTPHeaders = [
-            "Content-Type": "application/x-www-form-urlencoded",
-            "X-Naver-Client-Id": NaverApiHeaders.clientId,
-            "X-Naver-Client-Secret": NaverApiHeaders.clientSecret
-        ]
-        
-        let params: [String: Any] = [
-            "query": text
-        ]
-        
-        AF.request(url, method: .post, parameters: params, encoding: URLEncoding.default, headers: headers)
-            .validate(statusCode: 200..<201)
-            .responseDecodable(of: Sensing.self) { response in
-                switch response.result{
-                case .success (let data):
-                    completion(data.langCode)
-                case .failure (let error):
-                    print("Translate error!\n\(error)")
-                }
-            }
-    }
-    
     /// Codable 타입을 AnyObject 타입으로 변환
     /// - Parameter codableObject: AnyObject로 변환할 Codable 객체
     /// - Returns: AnyObject 타입의 객체
@@ -185,87 +135,5 @@ extension ChatService: StompClientLibDelegate {
         
         NotificationCenter.default.post(name: NotificationName.updateChatList, object: self)
         NotificationCenter.default.post(name: NotificationName.updateChatRoom, object: self)
-        NotificationCenter.default.post(name: NotificationName.updateTabBarIcon, object: self)
     }
 }
-
-
-// MARK: - Singleton Channel Model
-//class SingletonChannel {
-//    static let shared: SingletonChannel = SingletonChannel()
-//    private init() {}
-//    
-//    var list: [Channel] = [
-//        Channel(id: 1, name: "집좋아모임", messages: [
-//            Message(sender: Sender(userId: 1, userName: "제이디"), content: "Hello!", dateTime: "2023-08-23 13:41:44.889734", langCode: "en", isNew: false),
-//            Message(sender: Sender(userId: 1, userName: "제이디"), content: "Let's be friends~~", dateTime: "2023-08-23 13:41:53.889734", langCode: "en", isNew: false),
-//            Message(sender: Sender(userId: 2, userName: "HJ"), content: "반가워요!!", dateTime: "2023-08-23 13:43:13.889734", langCode: "ko", isNew: false),
-//            Message(sender: Sender(userId: 2, userName: "HJ"), content: "Nice to meet you!!", dateTime: "2023-08-23 13:44:44.889734", langCode: "ko", isNew: false),
-//            Message(sender: Sender(userId: 1, userName: "제이디"), content: "It's an honor to meet you like this!", dateTime: "2023-08-23 13:47:44.889734", langCode: "en", isNew: false),
-//            Message(sender: Sender(userId: 2, userName: "HJ"), content: "yeah, me too!!", dateTime: "2023-08-23 13:54:44.889734", langCode: "ko", isNew: false),
-//            Message(sender: Sender(userId: 3, userName: "노리"), content: "こんにちは! よろしくお願いします！！", dateTime: "2023-08-23 13:58:44.889734", langCode: "ja", isNew: true)
-//        ], userProfileImages: ["https://rf-aws-bucket.s3.ap-northeast-2.amazonaws.com/userDefault/defaultImage.jpg", "https://rf-aws-bucket.s3.ap-northeast-2.amazonaws.com/userDefault/defaultImage.jpg",
-//                               "https://rf-aws-bucket.s3.ap-northeast-2.amazonaws.com/userDefault/defaultImage.jpg"]),
-////        Channel(id: 2, name: "2번 모임", messages: [
-////
-////        ], userProfileImages: ["a", "a", "a"])
-//    ]
-//    
-//    /// 채널 리스트에서 특정 채널의 index 위치 가져오기
-//    /// - Parameter channelId: Channel ID
-//    /// - Returns: Index(Int)
-//    func getChannelIndex(_ channelId: Int) -> Int? {
-//        let index = list.firstIndex { $0.id == channelId }
-//        return index
-//    }
-//    
-//    /// 특정 채널의 메시지 리스트 가져오기
-//    /// - Parameter id: Channel ID
-//    /// - Returns: Messages
-//    func getChannelMessages(_ id: Int) -> [Message] {
-//        let index = list.firstIndex { $0.id == id }
-//        guard let index else { return [] }
-//        return list[index].messages
-//    }
-//    
-//    /// 특정 채널에 메시지 추가
-//    /// - Parameters:
-//    ///   - channelId: Channel ID
-//    ///   - message: Message
-//    func appendMessage(channelId: Int, message: Message) {
-//        let index = list.firstIndex { $0.id == channelId }
-//        guard let index else { return }
-//        list[index].messages.append(message)
-//    }
-//    
-//    func sortByLatest() {
-//        list.sort { $0.messages.last?.dateTime ?? "" > $1.messages.last?.dateTime ?? "" }
-//    }
-//    
-//    /// 메시지를 모두 읽음 여부로 설정 후 가장 첫 새로운 메시지의 Index를 반환
-//    /// - Parameter channelId: Channel ID
-//    /// - Returns: Index
-//    func readNewMessage(_ channelId: Int) -> Int {
-//        guard let index = getChannelIndex(channelId) else { return 0 }
-//        var messages = getChannelMessages(channelId)
-//        
-//        // 메시지 비어있으면 index = 0
-//        if messages.isEmpty { return 0 }
-//        
-//        var firstNewMessageIndex = messages.count
-//        
-//        // 가장 마지막 메시지가 새로운 메시지가 아닐 경우 마지막 index 리턴
-//        if messages.last?.isNew == false { return firstNewMessageIndex - 1}
-//        
-//        for i in stride(from: messages.count - 1, through: 0, by: -1) {
-//            if messages[i].isNew == false {
-//                break
-//            }
-//            messages[i].isNew = false
-//            firstNewMessageIndex -= 1
-//        }
-//        
-//        list[index].messages = messages
-//        return firstNewMessageIndex
-//    }
-//}
